# api/openapi.yaml
openapi: 3.0.3
info:
  title: Spacefleet API
  version: 0.1.0
  description: |
    HTTP API for the Spacefleet app. This is the source of truth: it
    generates the Go server stubs (lib/api/gen.go) and the TypeScript
    client types (ui/src/api/schema.d.ts). Run `make gen` after editing.

    Account management is organization-centric: organizations are the
    top-level tenant and a user must belong to (and have selected) one to use
    the app. /api/me bootstraps the SPA with the current user and their
    organization memberships.
servers:
  - url: /
paths:
...
  /api/organizations:
    post:
      summary: Create an organization (the caller becomes its first admin)
      operationId: createOrganization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationCreateRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
  /api/organizations/{id}:
    patch:
      summary: Rename an organization (admins only)
      operationId: renameOrganization
      parameters:
        - $ref: "#/components/parameters/OrganizationID"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OrganizationCreateRequest"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Organization"
        default:
          $ref: "#/components/responses/Error"
...
  /api/invitations:
    get:
      summary: List pending invitations for the current organization (admins only)
      description: Org-scoped (X-Organization-ID). Admin-only.
      operationId: listInvitations
      responses:
        "200":
          description: Pending invitations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Invitation"
        default:
          $ref: "#/components/responses/Error"
    post:
      summary: Invite a user to the current organization (admins only)
      description: |
        Org-scoped (X-Organization-ID). Admin-only. Creates a pending invitation
        and returns a copy-able accept link. When email is configured an
        invitation email is also sent (email_sent reflects whether it was
        enqueued); otherwise the admin shares the link manually.
      operationId: createInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/InvitationCreateRequest"
      responses:
        "201":
          description: Invitation created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InvitationCreateResult"
        default:
          $ref: "#/components/responses/Error"
...
components:
  schemas:
...
    Error:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
...
    Role:
      type: string
      description: |
        A member's role within an organization (hierarchical: admin > editor >
        viewer). admin manages the organization itself (members, invitations,
        renaming); editor can take any action within the app but not manage the
        org; viewer is read-only.
      enum: [admin, editor, viewer]
    Organization:
      type: object
      required: [id, name, created_at, updated_at]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    OrganizationCreateRequest:
      type: object
      required: [name]
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
...
    InvitationStatus:
      type: string
      enum: [pending, accepted, revoked]
    Invitation:
      type: object
      description: A pending (or resolved) invitation to join the organization.
      required: [id, email, role, status, expires_at, created_at, accept_url]
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
        role:
          $ref: "#/components/schemas/Role"
        status:
          $ref: "#/components/schemas/InvitationStatus"
        expires_at:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        accept_url:
          type: string
          description: The copy-able invite link an admin can share.
    InvitationCreateRequest:
      type: object
      required: [email, role]
      properties:
        email:
          type: string
          minLength: 1
          maxLength: 320
        role:
          $ref: "#/components/schemas/Role"
    InvitationCreateResult:
      type: object
      required: [invitation, email_sent]
      properties:
        invitation:
          $ref: "#/components/schemas/Invitation"
        email_sent:
          type: boolean
          description: |
            Whether an invitation email was enqueued. False when email isn't
            configured (or enqueueing failed) — share the accept_url manually.
...
  parameters:
    OrganizationID:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    Error:
      description: Error response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"